fix(server): Added correct js and css routing, error log
The js and css static files could not be correctly routed to just
using the route in the index.html that generated after build. To
fix this routing issue to the static files we explicitly define its
use in the server.js file. What a mess.

Changed the error log to be more descriptive about where the error
was caught